Crate bode_plot

A simple library for displaying bode plots in windows. Implement crate::plot::BodePlotTransferFunction for your own type and pass into create_plot or create_log_plot to display the frequency response of the transfer function.

Modules

plot

Constants

DEFAULT_HEIGHT

Default window height in pixels

DEFAULT_WIDTH

Default window width in pixels

Functions

create_log_plot

Asynchronously opens a window displaying the bode plot of the given transfer function with logarithmic scale.

create_plot

Asynchronously opens a window displaying the bode plot of the given transfer function with linear scale.